refactor(dht): move client to server file
