+++ src/toxcore/dht/mod.rs
@@ -27,4 +27,3 @@ pub mod kbucket;
 pub mod packed_node;
 pub mod codec;
 pub mod server;
-pub mod client;
